requirements.txt
opencv-python
numpy
matplotlib
ghostscript
pillow